changed

+++ teacher_bot/manage.py
@@ -1,5 +1,5 @@
 import logging
-from utils import BASE_URL,FeedBackTestNameStates, create_test_name, add_to_group, add_to_student, create_test, delete_group, delete_group_student, NameTestStates, FeedBackGroupNameStates,FeedBackStudentStates,GroupDeleteStates
+from utils import BASE_URL, add_to_group, add_to_student, create_test, delete_group, delete_group_student, NameTestStates, FeedBackGroupNameStates,FeedBackStudentStates,GroupDeleteStates
 from buttons import start_button, group_buttons, choose_group_button,group_button, group_test_buttons, add_button, delete_button, again_start_button
 from aiogram import Bot, Dispatcher, executor, types
 from aiogram.contrib.fsm_storage.memory import MemoryStorage
@@ -24,12 +24,14 @@ steps = {
     "begin": "begin",
     "enter_name": "enter_name",
     "group_add":"group_add",
-    "group_delete":"group_delet",
+    "group_delete":"group_delete",
+    "student_id_add":"student_id_add",
+    "enter_name_id":"enter_name_id",
     "student_delete":"student_delete",
     "student_deletes":"student_deletes",
+    "choose_group_students": "choose_group_students",
     "student_add":"student_add",
     "student_add_to_group":"student_add_to_group",
-    "group_delete":"group_delete",
     "group_view":"group_view",
     "enter_login": "enter_login",
     "choose": "choose",
@@ -48,6 +50,7 @@ data = {
         "test_name":"",
         "delete_groups_students":[],
         "test_names":[],
+        "student_id":"",
         'delete_group': [],
         'student_group': [],
         'group': [],
@@ -57,7 +60,7 @@ data = {
     },
 }
 
-bot = Bot(token=API_TOKEN)
+bot = Bot(token=API_TOKEN, parse_mode=types.ParseMode.HTML)
 
 dp = Dispatcher(bot, storage=MemoryStorage())
 
@@ -71,13 +74,30 @@ async def send_welcome(message: types.Message):
         reply_markup=start_button
     )
 
-
-        
 @dp.message_handler(Text(startswith="Boshlash"),lambda message: data["step"] == steps["begin"])
+async def begin(message: types.Message):
+        data["step"] = steps["enter_name_id"]
+        await message.answer("Iltimos loginingizni kiriting")
+        
+@dp.message_handler(lambda message: data["step"] == steps["enter_name_id"])
+async def begin(message: types.Message):
+    data["step"] = steps["enter_name_id"]
+    filtered_teacher = requests.get(url=f"{BASE_URL}/teachers/?user_id={message.text}")
+    teacher = filtered_teacher.json()
+    data["state"]["teacher"] = teacher
+    if teacher:
+        data["step"] = steps["enter_name"]
+        await message.answer("Iltimos ismingizni kiriting")
+    else:
+        data["step"] = steps["begin"]
+        await message.answer("Noto'g'ri login kiritdingiz.\nIltimos quyidagi tugmani bosing va boshqattan login kiriting", reply_markup=start_button)
+    
+
+
+@dp.message_handler(lambda message: data["step"] == steps["enter_name_id"])
 async def begin(message: types.Message):
     data["step"] = steps["enter_name"]
     await message.answer("Iltimos ismingizni kiriting")
-    
 
 
    
@@ -88,6 +108,31 @@ async def group_view(message: types.Message):
                         reply_markup=group_buttons(data["state"]["teacher"][0]['id']))
 
 
+@dp.message_handler(Text(startswith="O'quvchilar ro'yxatini ko'rish"))
+async def group_view(message: types.Message):
+    data["step"] = steps["choose_group_students"]
+    await message.answer("Quyidagilardan birini tanlang", 
+                        reply_markup=group_buttons(data["state"]["teacher"][0]['id']))
+    
+@dp.message_handler(lambda message: data["step"] == steps["choose_group_students"])
+async def group_delete(message: types.Message):
+    view_group_students = requests.get(url=f"{BASE_URL}/students/?group__name={message.text}").json()
+    if view_group_students:
+        view_students = f"<b>{message.text}</b> guruhdagi o'quvchilar ro'yxati:\n"
+        for view_student in view_group_students:
+            view_students += f"Ism:<b>{view_student['name']}</b> Login:<b>{view_student['user_id']}</b>\n"
+        data["step"] = steps["enter_name"]
+        await message.answer(view_students, reply_markup=choose_group_button)
+    else:
+        data["step"] = steps["enter_name"]
+        await message.answer("Guruhda hali o'quvchilar mavjud emas! Iltimos avval o'quvchi qo'shing", reply_markup=choose_group_button)
+
+
+
+
+
+    
+
 @dp.message_handler(Text(startswith="Guruhni yoki o'quvchini qo'shish"))
 async def group_or_student_add(message: types.Message):
     await message.answer("Quyidagilardan birin tanlang", reply_markup=add_button)
@@ -95,7 +140,7 @@ async def group_or_student_add(message: types.Message):
 
 @dp.message_handler(Text(startswith="Guruhni yoki o'quvchini o'chirish"))
 async def group_or_student_add(message: types.Message):
-    await message.answer("Quyidagilardan birin tanlang", reply_markup=delete_button)
+    await message.answer("Quyidagilardan birini tanlang", reply_markup=delete_button)
 
 
 @dp.message_handler(Text(startswith="Guruhni o'chirish"))
@@ -105,14 +150,19 @@ async def student_add(message: types.Message):
                         reply_markup=group_buttons(data["state"]["teacher"][0]['id']))
 
 @dp.message_handler(lambda message: data["step"] == steps["group_delete"])
-async def student_add(message: types.Message):
+async def group_delete(message: types.Message):
     deleted_group = requests.get(url=f"{BASE_URL}/groups/?name={message.text}").json()
     data["state"]["delete_group"] = deleted_group
-    grp_delete = delete_group(
-        data["state"]["delete_group"][0]['id']
-    )
-    response_text = f"{message.text}-guruh muvaffaqiyatli o'chirildi" if grp_delete else "Amal oxiriga yetmadi"
-    await message.answer(response_text)
+    if deleted_group:
+        grp_delete = delete_group(
+            data["state"]["delete_group"][0]['id']
+        )
+        data["step"] = steps["enter_name"]
+        await message.answer(f"<b>{message.text}</b>-guruh muvaffaqiyatli o'chirildi", reply_markup=choose_group_button)
+    else:
+        data["step"] = steps["enter_name"]
+        await message.answer("Amal oxiriga yetmadi", reply_markup=choose_group_button)
+        
  
 
 
@@ -130,15 +180,20 @@ async def student_delete_id(message: types.Message):
     data["step"] = steps["student_deletes"]
     await message.answer("O'chirmoqchi bo'lgan o'quvchingizni ismini kiriting")
 
+
 @dp.message_handler(lambda message: data["step"] == steps["student_deletes"])
 async def student_deletes_id(message: types.Message):
     student_deletes = data["state"]["delete_groups_students"][0]['name']
     filtered_group_student = requests.get(url=f"{BASE_URL}/students/?name={message.text}&group__name={student_deletes}").json()
-    respond = delete_group_student(
-        filtered_group_student[0]['id']
-    )
-    respond_text = f"{student_deletes} guruhdagi {message.text} ismli o'quvchi muvaffaqiyatli o'chirildi" if respond else "Amal oxiriga yetmadi"
-    await message.answer(respond_text)
+    if filtered_group_student:
+        respond = delete_group_student(
+            filtered_group_student[0]['id']
+        )
+        data["step"] = steps["enter_name"]
+        await message.answer(f"<b>{student_deletes}<b/> guruhdagi <b>{message.text}</b> ismli o'quvchi muvaffaqiyatli o'chirildi", reply_markup=choose_group_button)
+    else:
+        data["step"] = steps["enter_name"]
+        await message.answer("Bu guruhda bunday o'quvchi mavjud emas! Iltimos boshqa amalni bajarib ko'ring", reply_markup=choose_group_button)
 
 @dp.message_handler(Text(startswith="Guruhni qo'shish"))
 async def student_add(message: types.Message):
@@ -153,12 +208,20 @@ async def group_add_name(message: types.Message, state:FSMContext):
         message.text,
         data["state"]["teacher"][0]['id'],
     )
-    response_text = "Guruh yaratildi" if group_name else "Amal oxiriga yetmadi"
-    await message.answer(response_text)
-    await state.finish()
+    if group_name:
+        data["step"] = steps["enter_name"]
+        await message.answer("Guruh yaratildi", reply_markup=choose_group_button)
+        await state.finish()
+    else:
+        data["step"] = steps["enter_name"]
+        await message.answer("Amal oxiriga yetmadi", reply_markup=choose_group_button)
+
+    
+
+
 
 @dp.message_handler(Text(startswith="O'quvchini qo'shish"))
-async def student_add(message: types.Message):
+async def students_add(message: types.Message):
     data["step"] = steps["student_add_to_group"]
     await message.answer("Iltimos o'quvchi qo'shiladigan guruhni tanlang", reply_markup=group_buttons(data["state"]["teacher"][0]['id']))
 
@@ -167,79 +230,104 @@ async def student_add(message: types.Message):
 async def student_add(message: types.Message):
     filtered_group = requests.get(url=f"{BASE_URL}/groups/?name={message.text}").json()
     data["state"]["student_group"] = filtered_group
-    data["step"] = steps["student_add"]
+    data["step"] = steps["student_id_add"]
+    await message.answer("O'quvchini loginini kiriting")
+    # await FeedBackStudentStates.student_name.set()
 
+@dp.message_handler(lambda message: data["step"] == steps["student_id_add"])
+async def student_id_add(message: types.Message):
+    data["state"]["student_id"] = message.text
+    data["step"] = steps["student_add"]
     await message.answer("O'quvchini ismini kiriting")
     await FeedBackStudentStates.student_name.set()
 
 
 @dp.message_handler(lambda message: data["step"] == steps["student_add"], state=FeedBackStudentStates.student_name)
 async def student_add_name(message: types.Message, state:FSMContext):
-    print(data["state"]["student_group"])
     student_name = add_to_student(
         data["state"]["teacher"][0]['id'],
         data["state"]["student_group"][0]['id'],
         message.text,
+        data["state"]["student_id"],
     )
-    response_text = "O'quvchi guruhga kiritildi" if student_name else "Amal oxiriga yetmadi"
-    await message.answer(response_text)
-    await state.finish()
+    if student_name:
+        data["step"] = steps["enter_name"]
+        await message.answer("O'quvchi guruhga kiritildi", reply_markup=choose_group_button)
+        await state.finish()
+    else:
+        data["step"] = steps["enter_name"]
+        await message.answer("Amal oxiriga yetmadi", reply_markup=choose_group_button)
+
 
 
 @dp.message_handler(Text(startswith="Guruhga test jo'natish"))
 async def send_test(message: types.Message):
     data["step"] = steps["test_name_enter"]
     await message.answer("Iltimos testni nomini kiriting")
-    await FeedBackTestNameStates.name.set()
 
-@dp.message_handler(lambda message: data["step"] == steps["test_name_enter"],  state=FeedBackTestNameStates.name)
-async def choose_group(message: types.Message, state:FSMContext):
-    data["state"]["test_name"] = message.text
-    data["step"] = steps["test_name"]
-    test_name = create_test_name(
-        data["state"]["test_name"],
-    )
-    response_text = "Test kalitlarini kiriting" if test_name else "Amal oxiriga yetmadi"
-
-    await message.answer(response_text)
-    await state.finish()
-    await NameTestStates.test.set()
+@dp.message_handler(lambda message: data["step"] == steps["test_name_enter"])
+async def choose_group(message: types.Message):
+    filtered_tests = requests.get(url=f"{BASE_URL}/test_keys/?name={message.text}").json()
+    if filtered_tests:
+        data["step"] == steps["test_name_enter"]
+        await message.answer("Bu nomdan avval foydalanilgan.Iltimos testga boshqa nom bering:")
+    else:
+        data["state"]["test_name"] = message.text
+        data["step"] = steps["test_name"]
+        await message.answer("Iltimos testni kalitlarini kiriting:")
+        await NameTestStates.test.set()
+        
 
 @dp.message_handler(Text(startswith="Guruhni natijalarini bilish"))
 async def test_results(message: types.Message):
     data["step"] = steps["group_name_result"] 
-    await message.answer("Quyidagilardan birini tanlang", 
-                        reply_markup=group_test_buttons(data["state"]["group"][0]['id']))
+    guruh = data["state"]["groups"]
+    url = f"{BASE_URL}/test_keys/?group__name={guruh}"
+    responses = requests.get(url=url).json()
+    if responses:
+        await message.answer("Quyidagilardan birini tanlang", 
+                            reply_markup=group_test_buttons(data["state"]["group"][0]['id']))
+    else:
+        data["step"] = steps["choose_group"]
+        await message.answer("Ushbu guruhga hali test qo'shilmadi.Iltimos boshqa guruhni tanlang",
+                             reply_markup=group_buttons(data["state"]["teacher"][0]['id']))
 
 
 @dp.message_handler(Text(endswith="-Test"), lambda message: data["step"] == steps["group_name_result"])
 async def test_results(message: types.Message):
     filtered_tests = requests.get(url=f"{BASE_URL}/test_responses/?test__name={message.text}").json()
     gr = data["state"]["groups"]
-    result=f"{gr}-guruh {message.text} natijalari:\n"
-    for test_result in filtered_tests:
-        result += f"{test_result['student']['name']} {len(test_result['answer_message'])} ta testdan {test_result['correct_response_count']} topdi \n"
-    await message.answer(result, reply_markup=choose_group_button)
-    await message.answer("Boshqa amalni ko'ring", reply_markup=choose_group_button)
+    if filtered_tests:
+        result=f"{gr}-guruh {message.text} natijalari:\n"
+        for test_result in filtered_tests:
+            result += f"<b>{test_result['student']['name']}</b> <b>{len(test_result['answer_message'])}</b> ta testdan <b>{test_result['correct_response_count']}</b> topdi \n"
+        await message.answer(result)
+        data["step"] = steps["enter_name"]
+        await message.answer("Boshqa amalni ko'ring", reply_markup=choose_group_button)
+    else:
+        data["step"] = steps["enter_name"]
+        await message.answer("Hali natijalar mavjud emas!", reply_markup=choose_group_button)
 
 
 
 @dp.message_handler(lambda message: data["step"] == steps["test_name"], state=NameTestStates.test)
 async def create_tests(message: types.Message,  state:FSMContext):
     test_name = data["state"]["test_name"]
-    filtered_tests = requests.get(url=f"{BASE_URL}/tests/?name={test_name}").json()
     data["state"]["test_keys"] = message.text
-    await message.answer(message.text)
-
     result = create_test(
-        filtered_tests[0]['id'],
+        test_name,
         data["state"]["test_keys"],
         data["state"]["teacher"][0]['id'],
         data["state"]["group"][0]['id'],
     )
-    response_text = "Test yaratildi" if result else "Amal oxiriga yetmadi"
-    await message.answer(response_text)
-    await state.finish()
+    if result:
+        data["step"] = steps["enter_name"]
+        await message.answer("Test yaratildi", reply_markup=choose_group_button)
+        await state.finish()
+    else:
+        data["step"] = steps["enter_name"]
+        await message.answer("Amal oxiriga yetmadi", reply_markup=choose_group_button)
+
 
 
 
@@ -265,8 +353,8 @@ async def enter_name(message: types.Message):
             await message.answer("Quyidagilardan birini tanlang", 
                              reply_markup=choose_group_button)
     else:
-        data["step"] = steps["begin"]
-        await message.answer("Noto'g'ri ism kiritdingiz.\nIltimos quyidagi tugmani bosing va boshqattan ism kiriting", reply_markup=start_button)
+        data["step"] == steps["enter_name_id"]
+        await message.answer("Noto'g'ri ism kiritdingiz.\nIltimos quyidagi tugmani bosing va boshqattan ism kiriting")
 
 if __name__ == '__main__':
